Fixed some errors with linking files. Removed warnings with duplication of code, deperecations and dll-warnings. Hazel completely moved from dll to static library.

+++ Hazel/src/hazel/layer/imgui/ImGuiLayer.cpp
@@ -1,3 +1,5 @@
+#include "hzpch.h"
+
 #include "ImGuiBuilder.h"
 #include "ImGuiLayer.h"
 
@@ -67,7 +69,7 @@ namespace Hazel {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::getInstance();
-		io.DisplaySize = ImVec2(app.getWindow().getWidth(), app.getWindow().getHeight());
+		io.DisplaySize = ImVec2((float) app.getWindow().getWidth(),(float) app.getWindow().getHeight());
 		// Rendering...
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
